% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted-randomization.R
\name{restricted}
\alias{restricted}
\title{Restricted randomization}
\usage{
restricted(w, nsbj, procedure, parameter = NA, nsim = 1)
}
\arguments{
\item{w}{vector of fixed allocation ratio ((ir)rational or integers with GCD = 1).}

\item{nsbj}{number of subjects to randomize.}

\item{procedure}{name of randomization procedure \code{c("CRD", "PBD", "BUD", "MWUD",
"DBCD", "DL", "MinQD", "MaxEnt")}.}

\item{parameter}{tuning parameter of randomization procedure (NA for \code{"CRD"}).}

\item{nsim}{number of simulations (1 by default).}
}
\value{
In case of a single trial simulation (i.e. \code{nsim} equals to 1) returns
   a list with the following items items:
   \itemize{
      \item \code{op} -- a tibble (data frame) of operational characteristics:
      forcing index, imbalance.
      \item \code{probability} -- a tibble (data frame) of allocation probabilities at each step.
      \item \code{allocation} -- a tibble (data frame) of subjects' allocations at each step.
      \item \code{observation} -- a tibble (data frame) of treatment assignments.
   }

   In case of multiple trial simulations (i.e. \code{nsim} larger than 1) returns
   a list with the following items items:
   \itemize{
      \item \code{op} -- a tibble (data frame) of operational characteristics:
      average forcing index (AFI), maximum imbalance (MI), momentum of probability
      mass (MPM), cumulative momentum of probability mass (CMPM), and "average
      standard deviation" of the allocation proportions (ASD).
      \item \code{probability} -- a tibble (data frame) of unconditional allocation
      probabilities at each step.
      \item \code{allocation} -- a tibble (data frame) of distrbutions of the
      subjects' allocations at each step.
      \item \code{observation} -- a tibble (data frame) of distributions of treatment
      assignments.
   }
}
\description{
Simulates a clinical trial one time or multiple times with retricted randomizaion procedure
}
\examples{
   # a single trial simulation (without responses):
   #    target allocation ratio: 4:3:2:1 (i.e. 4 treatment groups)
   #    number of subjects involved in  trial: 100
   #    randomization procedure used: Drop-the-Loser (DL)
   #    randomization procedure parameter: 2

   restricted(c(4, 3, 2, 1), 200, "DL", 2)

   # 200 trial simulations without responses:
   #    target allocation ratio: 31:19 (i.e. 2 treatment groups)
   #    number of subjects involved in  trial: 50
   #    randomization procedure used: Maximum Entropy Constraint Balance
   #          Randomization (MaxEnt)
   #    randomization procedure parameter: 0.5
   #    number of simulations: 200

   restricted(c(31, 19), 50, "MaxEnt", 0.5, nsim = 200)

}
\references{
Ryeznik Y, Sverdlov O (2018) "A Comparative Study of Restricted Randomization Procedures
   for Multi-Arm Trials with Equal or Unequal Treatment Allocation Ratios",
   Statistics in Medicine, Submitted.
}
\author{
Yevgen Ryeznik (\email{yevgen.ryeznik@gmail.com}), Oleksandr Sverdlov
}
