% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-in-medicine-example.R
\name{run_example}
\alias{run_example}
\title{A simulation example}
\usage{
run_example(nsim = 10000)
}
\arguments{
\item{nsim}{number of simulations (by default equals to 10000)}
}
\value{
a list of simulations results per scenario
}
\description{
Reproduces simulations from the paper Ryeznik Y, Sverdlov O (2018) "A Comparative Study of Restricted Randomization
Procedures for Multi-Arm Trials with Equal or Unequal Treatment Allocation Ratios",
Statistics in Medicine, Submitted.
}
\examples{
   # 1000 simulations:

   run_example(nsim = 1000)

}
