% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-in-medicine-example.R
\name{visualize_example}
\alias{visualize_example}
\title{A simulation example: analysis and visualization}
\usage{
visualize_example(data_file = "./data/simulation_data_to_analyze.Rda")
}
\arguments{
\item{data_file}{a path to the *.Rda file where results of \code{run_example} are stored}
}
\description{
Visualize reproduced simulations from the paper Ryeznik Y, Sverdlov O (2018) "A Comparative Study of Restricted Randomization
Procedures for Multi-Arm Trials with Equal or Unequal Treatment Allocation Ratios",
Statistics in Medicine, Submitted.
}
\examples{
   # by default, run_example function creates a folder "data" in a working folder and stores the results in a files
   # "./data/simulation_data_to_analyze.Rda":

   visualize_example(data_file = "./data/simulation_data_to_analyze.Rda")

}
